Content management system for point-of-purchase display
This project was for a kiosk point-of-purchase downloadable content media player.  Basically this runs a kiosk display, in a retail store, that plays a looping attract video and responds to user input and plays the relevant video and activates appropriate LEDs.  It downloads both playlists and content from a central source so an operator can change the kiosk content remotely and easily.  Originally developed in Java running on a special Android™ distro, and ported to Python and running on Raspberry Pi™.

Using the same technology, packaged in another form factor to create a dynamic picture frame with marketing focus for seniors.