Convert time-lock safe report to searchable database
This project was to parse a human-readable report that is generated from a time-lock safe and convert it to an easier format that can be stored, retrievable, and searchable.  My part was to develop the PHP classes.  The safe was a time-locked type located in a gas station.  When the safe is opened by the armored truck service, the safe creates the report and sends it using ftp to a ftp server.