Translate Nucleus point-of-sale data to online format
This project was to take transaction data from an existing legacy point-of-sale system and send it to a web server.  The web server then stores the data and provides a user interface to view and search the point-of-sale transaction data.  The point-of-sale system uses the Nucleus POS protocol.  I worked on the Windows application that took the point-of-sale data and posted it onto the web server.  It consists of 2 parts, the POS application that minimized into the system tray, and the GUI that is used to configure the POS application.